+++ packages/transformer-react-styles-esm-imports/CHANGELOG.md
@@ -10,3 +10,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 ## 1.0.1-prerelease.0 (2023-11-02)
 
 **Note:** Version bump only for package transformer-react-styles-esm-imports
+
+## 1.0.1-prerelease.0 (2023-11-02)
+
+**Note:** Version bump only for package transformer-react-styles-esm-imports